requirements.txt
aiohttp==3.6.3
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
discord.py==1.5.1
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.6.0
gspread-asyncio==1.2.0
httplib2==0.18.1
idna==2.10
multidict==4.7.6
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
yarl==1.5.1
